requirements.txt
asgiref==3.8.1
basicauth==1.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==22.0.0
idna==3.7
packaging==24.0
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
sentry-sdk==2.0.1
sqlparse==0.5.0
stripe==9.4.0
typing_extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
